Emit empty structs instead of interface{} for empty types
When additionalProperties isn't specified, that means there aren't any
other meaningful values to be put into the type. Even though the json
schema would allow them to be specified, they won't be used. So we
don't need to allow them to be specified in the CRD types - we don't
need to add them as map[string]interface{}.

The deepcopy code generated by kubebuilder controller-gen for
map[string]interface{} doesn't compile; this allows us to get rid
of them.