Generate empty structs instead of interface{} for unknown types
These can be handled better by controller-gen, and it's probably more
correct for ARM - the API is unlikely to do anything useful with an
arbitrary type it gets sent.

We might have a pass to strip these out of the type graph. Although
maybe it's useful to keep them around for symmetry with the status
types (since they generally represent an object that only has
read-only properties in the swagger spec)?